use 64bit int where necessary

+++ list.h
@@ -1,4 +1,4 @@
-/* $Id: list.h,v 1.5 2004/01/31 21:31:55 aturner Exp $ */
+/* $Id: list.h,v 1.6 2004/04/03 22:42:20 aturner Exp $ */
 
 /*
  * Copyright (c) 2001-2004 Aaron Turner.
@@ -34,15 +34,15 @@
 #define _LIST_H_
 
 struct list_type {
-    int max;
-    int min;
+    u_int64_t max;
+    u_int64_t min;
     struct list_type *next;
 };
 
 typedef struct list_type LIST;
 
 int parse_list(LIST **, char *);
-int check_list(LIST *, int);
+int check_list(LIST *, u_int64_t);
 void free_list(LIST *);
 
 #endif
